% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdata}
\alias{cdata}
\title{Simulated competing risks data}
\format{
This data frame contains the following columns:

  \describe{

  \item{\code{ID}}{patient identifier.}
  \item{\code{surv}}{event time.}

  \item{\code{failure_type}}{event indicator. \code{0} denotes censoring, \code{1} risk 1, 
  and \code{2} risk 2.}

  \item{\code{x1}}{continuous variable.}

  \item{\code{x2}}{treatment indicator. \code{0} denotes the placebo group and \code{1} the treatment group.}

  }
}
\usage{
data(cdata)
}
\description{
The \code{cdata} data frame has 1000 rows and 5 columns.
}
\keyword{datasets}
