% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConcordanceJMMLSM.R
\name{ConcordanceJMMLSM}
\alias{ConcordanceJMMLSM}
\title{Concordance for joint models}
\usage{
ConcordanceJMMLSM(
  seed = 100,
  object,
  n.cv = 3,
  maxiter = 10000,
  initial.para = TRUE,
  ...
)
}
\arguments{
\item{seed}{a numeric value of seed to be specified for cross validation.}

\item{object}{object of class 'JMMLSM'.}

\item{n.cv}{number of folds for cross validation. Default is 3.}

\item{maxiter}{the maximum number of iterations of the EM algorithm that the 
function will perform. Default is 10000.}

\item{initial.para}{Initial guess of parameters for cross validation. Default is FALSE.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
a list of matrices with conditional probabilities for subjects.
}
\description{
Concordance for joint models
}
\seealso{
\code{\link{JMMLSM}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
