% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.jmcs.R
\name{fitted}
\alias{fitted}
\alias{fitted.jmcs}
\title{Fitted values for joint models}
\usage{
\method{fitted}{jmcs}(
  object,
  type = c("Marginal", "Subject"),
  process = c("Longitudinal", "Event"),
  ...
)
}
\arguments{
\item{object}{an object inheriting from class \code{jmcs}.}

\item{type}{for which type of fitted values to calculate.}

\item{process}{for which sub-model to calculate the fitted values.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Extract fitted values for joint models.
}
