% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mvcdata}
\alias{mvcdata}
\title{Simulated competing risks data correlated with mvydata}
\format{
This data frame contains the following columns:

  \describe{

  \item{\code{ID}}{patient identifier.}
  \item{\code{survtime}}{event time.}

  \item{\code{cmprsk}}{event indicator. \code{0} denotes censoring, \code{1} risk 1, 
  and \code{2} risk 2.}
  
  \item{\code{X21}}{X21.}
   
  \item{\code{X22}}{X22.}
  }
}
\usage{
data(mvcdata)
}
\description{
The \code{mvcdata} data frame has 500 rows and 5 columns.
}
\keyword{datasets}
