% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mvydata}
\alias{mvydata}
\title{Simulated bivariate longitudinal data}
\format{
This data frame contains the following columns:

  \describe{

  \item{\code{ID}}{patient identifier.}
  
  \item{\code{time}}{visit time.}
  
  \item{\code{Y1}}{response variable of biomarker 1.}
  
  \item{\code{Y2}}{response variable of biomarker 2.}

  \item{\code{X11}}{X11.}
   
   \item{\code{X12}}{X12.}
   
  }
}
\usage{
data(mvydata)
}
\description{
The \code{mvydata} data frame has 4060 rows and 6 columns.
}
\keyword{datasets}
