% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.survfitjmcs.R
\name{plot.survfitjmcs}
\alias{plot.survfitjmcs}
\title{Plot survival probabilities for new subjects}
\usage{
\method{plot}{survfitjmcs}(
  x,
  estimator = c("both", "mean", "median"),
  conf.int = TRUE,
  include.y = FALSE,
  ...
)
}
\arguments{
\item{x}{x of class \code{survfitjmcs}.}

\item{estimator}{character string specifying, whether to include in the plot 
the mean of the conditional probabilities of survival, the median or both. 
The mean and median are taken as estimates of these conditional probabilities 
over the M replications of the Monte Carlo scheme described in \code{survfitjmcs}.}

\item{conf.int}{logical; if TRUE, then a pointwise confidence interval is included in the plot. Default is TRUE.}

\item{include.y}{include longitudinal responses of this subject versus time. Default is FALSE.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Plot survival probabilities for new subjects.
}
