% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Brierjmcs.R
\name{print.Brierjmcs}
\alias{print.Brierjmcs}
\title{Print Brierjmcs}
\usage{
\method{print}{Brierjmcs}(x, ...)
}
\arguments{
\item{x}{x of class 'Brierjmcs'.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
a list of matrices with conditional probabilities for subjects.
}
\description{
Print Brierjmcs
}
\seealso{
\code{\link{jmcs}, \link{survfitjmcs}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
