% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.FastJM.R
\name{print.FastJM}
\alias{print.FastJM}
\title{Print FastJM}
\usage{
\method{print}{FastJM}(x, ...)
}
\arguments{
\item{x}{Object of class 'FastJM'.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print contents of FastJM object.
}
\seealso{
\code{\link{jmcs}}
}
\author{
Shanpeng Li
}
