% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.JMMLSM.R
\name{print.JMMLSM}
\alias{print.JMMLSM}
\title{Print JMMLSM}
\usage{
\method{print}{JMMLSM}(x, digits = 4, ...)
}
\arguments{
\item{x}{Object of class 'JMMLSM'.}

\item{digits}{number of digits of decimal to be printed.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
a summary of data, joint model, log likelihood, and parameter estimates.
}
\description{
Print contents of JMMLSM object.
}
\seealso{
\code{\link{JMMLSM}}
}
\author{
Shanpeng Li
}
