% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.JMMLSM.R, R/print.jmcs.R,
%   R/print.mvjmcs.R
\name{print.JMMLSM}
\alias{print.JMMLSM}
\alias{print}
\alias{print.jmcs}
\alias{print.mvjmcs}
\title{Print fitted joint models}
\usage{
\method{print}{JMMLSM}(x, digits = 4, ...)

\method{print}{jmcs}(x, digits = 4, ...)

\method{print}{mvjmcs}(x, digits = 4, ...)
}
\arguments{
\item{x}{A fitted model of class \code{jmcs},
\code{JMMLSM}, or \code{mvjmcs}.}

\item{digits}{Number of decimal places used when formatting
parameter estimates. Default is \code{4}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Print data summaries, model information, and
  parameter estimates from fitted joint models.
}
\seealso{
\code{\link{jmcs}}, \code{\link{JMMLSM}},
  \code{\link{mvjmcs}}
}
