% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.survfitJMMLSM.R, R/print.survfitjmcs.R,
%   R/print.survfitmvjmcs.R
\name{print.survfitJMMLSM}
\alias{print.survfitJMMLSM}
\alias{print.survfit}
\alias{print.survfitjmcs}
\alias{print.survfitmvjmcs}
\title{Print dynamic predictions from joint models}
\usage{
\method{print}{survfitJMMLSM}(x, ...)

\method{print}{survfitjmcs}(x, ...)

\method{print}{survfitmvjmcs}(x, ...)
}
\arguments{
\item{x}{A prediction object of class \code{survfitjmcs},
\code{survfitJMMLSM}, or \code{survfitmvjmcs}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Print subject-specific conditional survival
  probabilities for single-failure models or conditional
  cumulative incidence probabilities for competing-risks models.
}
\seealso{
\code{\link{survfitJM}}
}
