% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.survfitJMMLSM.R
\name{print.survfitJMMLSM}
\alias{print.survfitJMMLSM}
\title{Print survfitJMMLSM}
\usage{
\method{print}{survfitJMMLSM}(x, ...)
}
\arguments{
\item{x}{x of class 'survfitJMMLSM'.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
a list of matrices with conditional probabilities for subjects.
}
\description{
Print survfitJMMLSM
}
\seealso{
\code{\link{JMMLSM}, \link{survfitJMMLSM}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
