% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.survfitmvjmcs.R
\name{print.survfitmvjmcs}
\alias{print.survfitmvjmcs}
\title{Print survfitmvjmcs}
\usage{
\method{print}{survfitmvjmcs}(x, ...)
}
\arguments{
\item{x}{x of class 'survfitmvjmcs'.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
a list of matrices with conditional probabilities for subjects.
}
\description{
Print survfitmvjmcs
}
\seealso{
\code{\link{mvjmcs}, \link{survfitmvjmcs}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
