% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.jmcs.R
\name{residuals}
\alias{residuals}
\alias{residuals.jmcs}
\title{Residuals for joint models}
\usage{
\method{residuals}{jmcs}(object, type = c("Marginal", "Subject"), ...)
}
\arguments{
\item{object}{an object inheriting from class \code{jmcs}.}

\item{type}{what type of residuals to calculate.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Extract residuals for joint models.
}
