% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SimDataSF}
\alias{SimDataSF}
\title{Data simulation of continuous outcomes and survival data with single failure}
\usage{
SimDataSF(
  k_val,
  p1_val,
  p1a_val,
  p2_val,
  g_val,
  truebeta,
  truegamma,
  truevee,
  randeffect,
  yfn,
  cfn,
  mfn
)
}
\arguments{
\item{k_val}{The number of subjects in study.}

\item{p1_val}{The dimension of fixed effects in longitudinal measurements.}

\item{p1a_val}{The dimension of random effects in longitudinal measurements.}

\item{p2_val}{The dimension of fixed effects in competing risks failure time data.}

\item{g_val}{The number of type of failure in survival data. It must be 1.}

\item{truebeta}{True values for beta, the longitudinal coefficients.}

\item{truegamma}{True values for gamma, the survival coefficients.}

\item{truevee}{True values for \eqn{\nu_1}.}

\item{randeffect}{True values for random effects in longitudinal and survival sub-models.}

\item{yfn}{Filename of genereated Y matrix for longitudinal measurements in long format.}

\item{cfn}{Filename of genereated C matrix for competing risks failure time data.}

\item{mfn}{Filename of genereated M vector to indicate the number of longitudinal measurements per subject.}
}
\value{
Files with names yfn, cfn and mfn.
\tabular{ll}{
\code{censoring_rate} \tab Censoring rate of the survival data. \cr
\code{rate} \tab Rate of the event. \cr
\code{yfn} \tab Filename of genereated Y matrix for longitudinal measurements. \cr
\code{cfn} \tab Filename of genereated C matrix for competing risks failure time data. \cr
\code{mfn} \tab  Filename of genereated M vector to indicate the number of longitudinal measurements per subject.
  }
}
\description{
Data simulation of continuous outcomes and survival data with single failure

Data simulation of continuous outcomes and competing risks
}
\examples{
# A toy example testint data generations
require(FastJM)
set.seed(100)
yfn=tempfile(pattern = "", fileext = ".txt")
cfn=tempfile(pattern = "", fileext = ".txt")
mfn=tempfile(pattern = "", fileext = ".txt")
k_val=100;
p1_val=3;
p1a_val=2;
p2_val=2;
g_val=1
truegamma=c(0.8,-1)
truebeta=c(2,1,-1.5);
randeffect=c(0.5,0.5,0.25);
truevee1=c(1,0.5);
truevee2=c(0.7,0.25);
#writing files
SimDataSF(k_val, p1_val, p1a_val, p2_val, g_val,truebeta,
        truegamma, truevee1, randeffect, yfn,  cfn,  mfn)
\donttest{
fit <- jmcs(p1 = 3, yfile = yfn, cfile = cfn, mfile = mfn,
type_file = T, point = 6, do.trace = F)
}
}
