% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simJMdata.R
\name{simJMdata}
\alias{simJMdata}
\title{Simulate single-biomarker joint model data}
\usage{
simJMdata(
  seed = 100,
  N = 200,
  increment = 0.7,
  beta = c(5, 1.5, 2, 1, 2),
  sigma2 = 1,
  gamma1 = c(1, 0.5, 0.5),
  gamma2 = c(-0.5, 0.5, 0.25),
  alpha1 = c(1, 0.7),
  alpha2 = c(-1, -0.5),
  lambda1 = 0.05,
  lambda2 = 0.025,
  CL = 5,
  CU = 10,
  covb = diag(rep(1, 2)),
  CR = TRUE
)
}
\arguments{
\item{seed}{Integer random seed used for data generation. Default is
\code{100}.}

\item{N}{Integer specifying the sample size. Default is \code{200}.}

\item{increment}{Numeric value specifying the increment of visit times for
longitudinal measurements. Default is \code{0.7}.}

\item{beta}{Numeric vector of true fixed-effect parameters for the
longitudinal sub-model.}

\item{sigma2}{Scalar parameter of error variance for the
longitudinal sub-model.}

\item{gamma1}{Numeric vector of true fixed-effect parameters in the
cause-specific hazard sub-model for failure type 1.}

\item{gamma2}{Numeric vector of true fixed-effect parameters in the
cause-specific hazard sub-model for failure type 2.}

\item{alpha1}{Numeric vector of association parameters of individual mean linking the
longitudinal process to the cause-specific hazard for failure type 1.}

\item{alpha2}{Numeric vector of association parameters individual mean linking the
longitudinal process to the cause-specific hazard for failure type 2.}

\item{lambda1}{Baseline hazard rate for failure type 1. An exponential
baseline hazard with rate \code{lambda1} is assumed.}

\item{lambda2}{Baseline hazard rate for failure type 2. An exponential
baseline hazard with rate \code{lambda2} is assumed.}

\item{CL}{Lower bound of the uniform distribution used to generate censoring
times.}

\item{CU}{Upper bound of the uniform distribution used to generate censoring
times.}

\item{covb}{Variance-covariance matrix for the random effects in the
longitudinal sub-model.}

\item{CR}{Logical; if \code{TRUE}, competing risks data with two failure
types are simulated. If \code{FALSE}, a single-failure time-to-event
outcome is generated. Default is \code{TRUE}.}
}
\value{
A list with the following components:
\describe{
  \item{\code{ydata}}{A long-format data frame containing the simulated
  longitudinal biomarker measurements.}
  \item{\code{cdata}}{A data frame containing the simulated event-time data.}
}
}
\description{
Simulates single longitudinal biomarker data and associated time-to-event data. 
The survival outcome may be generated either as a single-failure time-to-event outcome or 
as competing risks data with two failure types.
}
\details{
Simulate single-biomarker joint model data
}
