% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.AUCjmcs.R
\name{summary.AUCjmcs}
\alias{summary.AUCjmcs}
\title{Print AUCjmcs}
\usage{
\method{summary}{AUCjmcs}(object, digits = 4, ...)
}
\arguments{
\item{object}{object of class 'AUCjmcs'.}

\item{digits}{number of digits of decimal to be printed.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
a list of matrices with conditional probabilities for subjects.
}
\description{
Print AUCjmcs
}
\seealso{
\code{\link{jmcs}, \link{survfitjmcs}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
