% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.AUCJMMLSM.R
\name{summary.AUCJMMLSM}
\alias{summary.AUCJMMLSM}
\title{Print AUCJMMLSM}
\usage{
\method{summary}{AUCJMMLSM}(object, digits = 4, ...)
}
\arguments{
\item{object}{object of class 'AUCJMMLSM'.}

\item{digits}{number of digits of decimal to be printed.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
a list of matrices with conditional probabilities for subjects.
}
\description{
Print AUCJMMLSM
}
\seealso{
\code{\link{JMMLSM}, \link{survfitJMMLSM}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
