% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.Concordance.R
\name{summary.Concordance}
\alias{summary.Concordance}
\title{Summarize concordance estimates}
\usage{
\method{summary}{Concordance}(object, digits = 4, ...)
}
\arguments{
\item{object}{An object of class \code{Concordance}.}

\item{digits}{Number of decimal places used for rounding.
Default is \code{4}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A data frame containing the averaged concordance estimates.
}
\description{
Summarize concordance estimates averaged across
  cross-validation folds.
}
\seealso{
\code{\link{Concordance}}
}
