% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.Concordancejmcs.R
\name{summary.Concordancejmcs}
\alias{summary.Concordancejmcs}
\title{Print Concordancejmcs}
\usage{
\method{summary}{Concordancejmcs}(object, digits = 4, ...)
}
\arguments{
\item{object}{object of class 'Concordancejmcs'.}

\item{digits}{number of digits of decimal to be printed.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
a list of matrices with conditional probabilities for subjects.
}
\description{
Print Concordancejmcs
}
\seealso{
\code{\link{jmcs}, \link{survfitjmcs}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
