% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ConcordanceJMMLSM.R
\name{summary.ConcordanceJMMLSM}
\alias{summary.ConcordanceJMMLSM}
\title{Print ConcordanceJMMLSM}
\usage{
\method{summary}{ConcordanceJMMLSM}(object, digits = 4, ...)
}
\arguments{
\item{object}{object of class 'ConcordanceJMMLSM'.}

\item{digits}{number of digits of decimal to be printed.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
a list of matrices with conditional probabilities for subjects.
}
\description{
Print ConcordanceJMMLSM
}
\seealso{
\code{\link{JMMLSM}, \link{survfitJMMLSM}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
