% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.DynPredAcc.R
\name{summary.DynPredAcc}
\alias{summary.DynPredAcc}
\title{Summarize dynamic prediction accuracy}
\usage{
\method{summary}{DynPredAcc}(
  object,
  metric = c("AUC", "Cindex", "Brier Score", "MAE", "MAEQ"),
  digits = 4,
  ...
)
}
\arguments{
\item{object}{An object of class \code{DynPredAcc}.}

\item{metric}{A character string specifying the metric:
\code{"AUC"}, \code{"Cindex"}, \code{"Brier Score"},
\code{"MAE"}, or \code{"MAEQ"}.
The metric must have been computed in \code{object}.
Default is \code{"AUC"}.}

\item{digits}{Number of decimal places used for rounding.
Default is \code{4}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A data frame containing the selected metric averaged
  across cross-validation folds at the prediction horizons.
}
\description{
Summarize prediction accuracy estimates averaged
  across cross-validation folds.
}
\seealso{
\code{\link{DynPredAcc}}
}
