% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.MAEQjmcs.R
\name{summary.MAEQjmcs}
\alias{summary.MAEQjmcs}
\title{Print MAEQjmcs}
\usage{
\method{summary}{MAEQjmcs}(x, ...)
}
\arguments{
\item{x}{x of class 'MAEQjmcs'.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
a list of matrices with conditional probabilities for subjects.
}
\description{
Print MAEQjmcs
}
\seealso{
\code{\link{jmcs}, \link{survfitjmcs}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
