% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.MAEQJMMLSM.R
\name{summary.MAEQJMMLSM}
\alias{summary.MAEQJMMLSM}
\title{Print MAEQJMMLSM}
\usage{
\method{summary}{MAEQJMMLSM}(object, digits = 3, ...)
}
\arguments{
\item{object}{object of class 'MAEQJMMLSM'.}

\item{digits}{number of decimal points to be rounded.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
a list of matrices with conditional probabilities for subjects.
}
\description{
Print MAEQJMMLSM
}
\seealso{
\code{\link{JMMLSM}, \link{survfitJMMLSM}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
