% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.PAjmcs.R
\name{summary.PAjmcs}
\alias{summary.PAjmcs}
\title{Print PAjmcs}
\usage{
\method{summary}{PAjmcs}(object, digits = 3, ...)
}
\arguments{
\item{object}{object of class 'PAjmcs'.}

\item{digits}{number of decimal points to be rounded.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
a list of matrices with conditional probabilities for subjects.
}
\description{
Print PAjmcs
}
\seealso{
\code{\link{jmcs}, \link{survfitjmcs}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
