% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.PEjmcs.R
\name{summary.PEjmcs}
\alias{summary.PEjmcs}
\title{Print PEjmcs}
\usage{
\method{summary}{PEjmcs}(object, error = c("MAE", "Brier"), ...)
}
\arguments{
\item{object}{object of class 'PEjmcs'.}

\item{error}{a character string that specifies the loss function.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
a list of matrices with conditional probabilities for subjects.
}
\description{
Print PEjmcs
}
\seealso{
\code{\link{jmcs}, \link{survfitjmcs}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
