% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.PEJMMLSM.R
\name{summary.PEJMMLSM}
\alias{summary.PEJMMLSM}
\title{Print PEJMMLSM}
\usage{
\method{summary}{PEJMMLSM}(object, error = c("MAE", "Brier"), ...)
}
\arguments{
\item{object}{object of class 'PEJMMLSM'.}

\item{error}{a character string that specifies the loss function.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
a list of matrices with conditional probabilities for subjects.
}
\description{
Print PEJMMLSM
}
\seealso{
\code{\link{JMMLSM}, \link{survfitJMMLSM}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
