% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survPAjmcs.R
\name{survPAjmcs}
\alias{survPAjmcs}
\title{Prediction in Joint Models}
\usage{
survPAjmcs(
  object,
  ynewdata = NULL,
  cnewdata = NULL,
  u = NULL,
  s = NULL,
  obs.time = NULL,
  LOCF = FALSE,
  LOCFcovariate = NULL,
  clongdata = NULL,
  quadpoint = NULL,
  ...
)
}
\arguments{
\item{object}{an object inheriting from class \code{jmcs}.}

\item{ynewdata}{a data frame that contains the longitudinal and covariate information for the subjects 
for which prediction of survival probabilities is required.}

\item{cnewdata}{a data frame that contains the survival and covariate information for the subjects 
for which prediction of survival probabilities is required.}

\item{u}{a numeric vector of times for which prediction survival probabilities are to be computed.}

\item{s}{a numeric saclar. This specifies the known time at which each of 
the subjects in cnewdata was known to be alive.}

\item{obs.time}{a character string of specifying a longitudinal time variable in ynewdata.}

\item{LOCF}{a logical value to indicate whether the last-observation-carried-forward approach applies to prediction. 
If \code{TRUE}, then \code{LOCFcovariate} and \code{clongdata} must be specified to indicate 
which time-dependent survival covariates are included for dynamic prediction. Default is FALSE.}

\item{LOCFcovariate}{a vector of string with time-dependent survival covariates if \code{LOCF = TRUE}. Default is NULL.}

\item{clongdata}{a long format data frame where time-dependent survival covariates are incorporated. Default is NULL.}

\item{quadpoint}{number of quadrature points used for estimating conditional probabilities.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a list of matrices with conditional probabilities for subjects.
}
\description{
This function computes R square.
}
\seealso{
\code{\link{jmcs}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
