% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.JMMLSM.R, R/vcov.jmcs.R, R/vcov.mvjmcs.R
\name{vcov.JMMLSM}
\alias{vcov.JMMLSM}
\alias{vcov}
\alias{vcov.jmcs}
\alias{vcov.mvjmcs}
\title{Extract variance-covariance matrices from fitted joint models}
\usage{
\method{vcov}{JMMLSM}(object, ...)

\method{vcov}{jmcs}(object, ...)

\method{vcov}{mvjmcs}(object, ...)
}
\arguments{
\item{object}{A fitted model of class \code{jmcs},
\code{JMMLSM}, or \code{mvjmcs}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A data frame containing the variance-covariance matrix
  of the parameter estimates, with rows and columns named
  by parameter.
}
\description{
Extract the variance-covariance matrix of estimated
  parameters from a fitted joint model.
}
\seealso{
\code{\link{jmcs}}, \code{\link{JMMLSM}},
  \code{\link{mvjmcs}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
