% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ydata}
\alias{ydata}
\title{Simulated longitudinal data}
\format{
This data frame contains the following columns:

  \describe{

  \item{\code{ID}}{patient identifier.}

  \item{\code{response}}{response variable.}

  \item{\code{time}}{visit time.}

  \item{\code{x1}}{treatment indicator. \code{0} denotes the placebo group and \code{1} the treatment group.}

  }
}
\usage{
data(ydata)
}
\description{
The \code{ydata} data frame has 3067 rows and 4 columns.
}
\keyword{datasets}
