% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ydatah}
\alias{ydatah}
\title{Simulated longitudinal data with within-subject variance}
\format{
This data frame contains the following columns:

  \describe{

  \item{\code{ID}}{patient identifier.}

  \item{\code{Y}}{response variable.}

  \item{\code{time}}{visit time.}

  \item{\code{Z1}}{treatment indicator. \code{0} denotes the placebo group and \code{1} the treatment group.}
   
   \item{\code{Z2}}{continuous variable..}
   
   \item{\code{Z3}}{continuous variable..}
  }
}
\usage{
data(ydatah)
}
\description{
The \code{ydatah} data frame has 1353 rows and 6 columns.
}
\keyword{datasets}
