% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixef.jmcs.R
\name{fixef}
\alias{fixef}
\alias{fixef.jmcs}
\title{Estimated coefficients estimates for joint models}
\usage{
\method{fixef}{jmcs}(object, process = c("Longitudinal", "Event"), ...)
}
\arguments{
\item{object}{an object inheriting from class \code{jmcs}.}

\item{process}{for which sub-model to extract the estimated coefficients.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Extracts the fixed effects for a fitted joint model.
}
