% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmcs_control.R
\name{jmcs_control}
\alias{jmcs_control}
\title{Control Options for jmcs}
\usage{
jmcs_control(
  quadpoint = NULL,
  maxiter = 10000,
  verbose = FALSE,
  initial.para = NULL,
  tol = 1e-04,
  method = c("pseudo-adaptive", "standard"),
  opt = c("nlminb", "optim")
)
}
\arguments{
\item{quadpoint}{Number of Gauss--Hermite quadrature points used for numerical
integration. The default is \code{NULL}.}

\item{maxiter}{Maximum number of EM iterations. The default is \code{10000}.}

\item{verbose}{Logical value indicating whether to print iteration details.
The default is \code{FALSE}.}

\item{initial.para}{Optional list of user-supplied initial parameter values
for the EM algorithm. The default is \code{NULL}.}

\item{tol}{Convergence tolerance. The default is \code{1e-4}.}

\item{method}{Numerical integration method used in the E-step. Available
options are \code{"pseudo-adaptive"} and \code{"standard"}. The default is
\code{"pseudo-adaptive"}.}

\item{opt}{Optimization method used to fit the initial linear mixed-effects
model. Available options are \code{"nlminb"} and \code{"optim"}. The default
is \code{"nlminb"}.}
}
\value{
A list of control parameters used by \code{\link{jmcs}()}.
}
\description{
Control options for \code{jmcs()}
}
\details{
Specifies numerical integration, convergence, initialization, and optimizer
settings for \code{\link{jmcs}()}.
}
