% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JMMLSM_control.R
\name{JMMLSM_control}
\alias{JMMLSM_control}
\title{Control Options for JMMLSM}
\usage{
JMMLSM_control(
  maxiter = 1000,
  tol = 1e-04,
  quadpoint = NULL,
  verbose = FALSE,
  initial.para = NULL,
  method = c("adaptive", "standard"),
  opt = c("nlminb", "optim")
)
}
\arguments{
\item{maxiter}{Maximum number of EM iterations. The default is \code{1000}.}

\item{tol}{Convergence tolerance. The default is \code{1e-4}.}

\item{quadpoint}{Number of Gauss--Hermite quadrature points. The default is
\code{15}.}

\item{verbose}{Logical value indicating whether to print iteration details.
The default is \code{FALSE}.}

\item{initial.para}{Optional list of user-supplied initial parameter values.
The default is \code{NULL}.}

\item{method}{Numerical integration method used in the E-step. Available
options are \code{"adaptive"} and \code{"standard"}. The default is
\code{"adaptive"}.}

\item{opt}{Optimization method used to fit the initial linear mixed-effects
model. Available options are \code{"nlminb"} and \code{"optim"}. The default
is \code{"nlminb"}.}
}
\value{
A list of control parameters used by \code{\link{JMMLSM}()}.
}
\description{
Control options for \code{JMMLSM()}
}
\details{
Specifies numerical integration, convergence, initialization, and optimizer
settings for \code{\link{JMMLSM}()}.
}
