% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.jmcs.R
\name{plot.jmcs}
\alias{plot.jmcs}
\title{Fitted values for joint models}
\usage{
\method{plot}{jmcs}(x, add.smooth = getOption("add.smooth"), ...)
}
\arguments{
\item{x}{x of class 'jmcs'.}

\item{add.smooth}{logical; if \code{TRUE} a smooth line is superimposed in the "Residuals vs Fitted" plot.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Plot Diagnostics for Joint Models.
}
