% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.survfitJMMLSM.R, R/plot.survfitjmcs.R,
%   R/plot.survfitmvjmcs.R
\name{plot}
\alias{plot}
\alias{plot.survfitJMMLSM}
\alias{plot.survfitjmcs}
\alias{plot.survfitmvjmcs}
\title{Plot conditional probabilities for new subjects}
\usage{
\method{plot}{survfitJMMLSM}(
  x,
  include.y = FALSE,
  xlab = NULL,
  ylab = NULL,
  xlim = NULL,
  ylim.long = NULL,
  ylim.surv = NULL,
  ...
)

\method{plot}{survfitjmcs}(
  x,
  include.y = FALSE,
  xlab = NULL,
  ylab = NULL,
  xlim = NULL,
  ylim.long = NULL,
  ylim.surv = NULL,
  ...
)

\method{plot}{survfitmvjmcs}(x, subject = NULL, risk = 1, include.y = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{survfitmvjmcs}.}

\item{include.y}{Logical; retained for consistency with other plotting
methods. The current method always displays longitudinal biomarker panels.
Default is \code{TRUE}.}

\item{xlab}{X axis label.}

\item{ylab}{Y axis label.}

\item{xlim}{X axis support.}

\item{ylim.long}{Y axis support for the longitudinal outcome.}

\item{ylim.surv}{Y axis support for the event / survival probability.}

\item{...}{Additional graphical arguments passed to the biomarker point plot.}

\item{subject}{Optional subject or subjects to plot. Can be a subject ID,
a numeric subject index, or a vector of IDs or indices. If \code{NULL}, all
subjects in \code{x$Last.time} are plotted. Default is \code{NULL}.}

\item{risk}{Failure type to plot when \code{x$CompetingRisk = TRUE}.
Default is \code{1}.}
}
\value{
plots of conditional probabilities over different pre-specified time points for subjects. 
If single failure type, then survival probabilities will be returned. 
Otherwise, cumulative incidence probabilities for each failure type will be returned.

plots of conditional probabilities over different pre-specified time points for subjects. 
If single failure type, then survival probabilities will be returned. 
Otherwise, cumulative incidence probabilities for each failure type will be returned.

Invisibly returns \code{x}. The function is called for its plotting
side effect.
}
\description{
Plot conditional probabilities for new subjects. 
If \code{CompetingRisk = FALSE}, print the survival probabilities. 
Otherwise, print the cumulative incidence probabilities for each failure type.

Plot conditional probabilities for new subjects. 
If \code{CompetingRisk = FALSE}, print the survival probabilities. 
Otherwise, print the cumulative incidence probabilities for each failure type.

Plot subject-specific conditional survival probabilities or cumulative
incidence probabilities from a \code{survfitmvjmcs} object. Longitudinal
observations for multiple biomarkers are stacked vertically, with the
predicted survival or cumulative incidence curve overlaid on each panel.
}
\seealso{
\code{\link{survfitJMMLSM}}

\code{\link{survfitjmcs}}

\code{\link{survfitmvjmcs}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
