% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.jmcs.R, R/print.mvjmcs.R
\name{print}
\alias{print}
\alias{print.jmcs}
\alias{print.mvjmcs}
\title{Print jmcs}
\usage{
\method{print}{jmcs}(x, digits = 4, ...)

\method{print}{mvjmcs}(x, digits = 4, ...)
}
\arguments{
\item{x}{Object of class 'mvjmcs'.}

\item{digits}{the number of significant digits to use when printing.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
a summary of data, joint model, log likelihood, and parameter estimates.

a summary of data, joint model, log likelihood, and parameter estimates.
}
\description{
Print jmcs

Print mvjmcs
}
\seealso{
\code{\link{jmcs}}

\code{\link{mvjmcs}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
