% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.jmcs.R
\name{print}
\alias{print}
\alias{print.jmcs}
\title{Print jmcs}
\usage{
\method{print}{jmcs}(x, digits = 4, ...)
}
\arguments{
\item{x}{Object of class 'jmcs'.}

\item{digits}{the number of significant digits to use when printing.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
a summary of data, joint model, log likelihood, and parameter estimates.
}
\description{
Print jmcs
}
\seealso{
\code{\link{jmcs}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
