% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.survfitjmcs.R
\name{print.survfitjmcs}
\alias{print.survfitjmcs}
\title{Print survfitjmcs}
\usage{
\method{print}{survfitjmcs}(x, ...)
}
\arguments{
\item{x}{x of class 'survfitjmcs'.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
a list of matrices with conditional probabilities for subjects.
}
\description{
Print survfitjmcs
}
\seealso{
\code{\link{jmcs}, \link{survfitjmcs}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
