% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.R
\name{ranef}
\alias{ranef}
\title{Random effects estimates for joint models}
\usage{
ranef(object, ...)
}
\arguments{
\item{object}{an object inheriting from class \code{jmcs}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Extracts the posterior mean of the random effects for a fitted joint model.
}
