% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.Concordance.R
\name{summary.Concordance}
\alias{summary.Concordance}
\title{Summarize Concordance}
\usage{
\method{summary}{Concordance}(object, digits = 4, ...)
}
\arguments{
\item{object}{An object of class 'Concordance'.}

\item{digits}{Number of decimal places to be printed. Default is \code{4}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A data frame containing the cross-validation averaged concordance estimate.
}
\description{
Summarize Concordance
}
\seealso{
\code{\link{Concordance}}, \code{\link{jmcs}}, \code{\link{JMMLSM}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
