% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.MAEQjmcs.R
\name{summary.MAEQjmcs}
\alias{summary.MAEQjmcs}
\title{Print MAEQjmcs}
\usage{
\method{summary}{MAEQjmcs}(object, digits = 3, ...)
}
\arguments{
\item{object}{object of class 'MAEQjmcs'.}

\item{digits}{number of decimal points to be rounded.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
a list of matrices with conditional probabilities for subjects.
}
\description{
Print MAEQjmcs
}
\seealso{
\code{\link{jmcs}, \link{survfitjmcs}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
