% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.DynPredAcc.R, R/summary.JMMLSM.R,
%   R/summary.jmcs.R, R/summary.mvjmcs.R
\name{summary}
\alias{summary}
\alias{summary.DynPredAcc}
\alias{summary.JMMLSM}
\alias{summary.jmcs}
\alias{summary.mvjmcs}
\title{Summaries of evaluation metrics for joint models}
\usage{
\method{summary}{DynPredAcc}(
  object,
  metric = c("AUC", "Cindex", "Brier Score", "MAE", "MAEQ"),
  digits = 4,
  ...
)

\method{summary}{JMMLSM}(object, process = c("longitudinal", "Event"), digits = 4, ...)

\method{summary}{jmcs}(object, process = c("Longitudinal", "Event"), digits = 4, ...)

\method{summary}{mvjmcs}(object, process = c("Longitudinal", "Event"), digits = 4, ...)
}
\arguments{
\item{object}{an object inheriting from class \code{mvjmcs}.}

\item{metric}{a list to indicate what metric to summarize}

\item{digits}{the number of significant digits to use when printing. Default is 4.}

\item{...}{further arguments passed to or from other methods.}

\item{process}{for which model (i.e., longitudinal model or survival model) to extract the estimated coefficients.}
}
\value{
a summary of the list of matrices with conditional probabilities for subjects.

A table to summarize the model results.

A table to summarize the model results.

A table to summarize the model results.
}
\description{
Produce result summaries of a joint model fit.

Produce result summaries of a joint model fit.

Produce result summaries of a joint model fit.
}
\seealso{
\code{\link{JMMLSM}}

\code{\link{jmcs}}

\code{\link{mvjmcs}}
}
