% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ConcordanceJMMLSM.R,
%   R/summary.Concordancejmcs.R, R/summary.DynPredAccJMMLSM.R,
%   R/summary.DynPredAccjmcs.R, R/summary.JMMLSM.R, R/summary.jmcs.R,
%   R/summary.mvjmcs.R
\name{summary}
\alias{summary}
\alias{summary.ConcordanceJMMLSM}
\alias{summary.Concordancejmcs}
\alias{summary.DynPredAccJMMLSM}
\alias{summary.DynPredAccjmcs}
\alias{summary.JMMLSM}
\alias{summary.jmcs}
\alias{summary.mvjmcs}
\title{Print ConcordanceJMMLSM}
\usage{
\method{summary}{ConcordanceJMMLSM}(object, digits = 4, ...)

\method{summary}{Concordancejmcs}(object, digits = 4, ...)

\method{summary}{DynPredAccJMMLSM}(
  object,
  metric = c("AUC", "Cindex", "Brier", "MAE", "MAEQ"),
  digits = 4,
  ...
)

\method{summary}{DynPredAccjmcs}(
  object,
  metric = c("AUC", "Cindex", "Brier", "MAE", "MAEQ"),
  digits = 4,
  ...
)

\method{summary}{JMMLSM}(object, process = c("longitudinal", "Event"), digits = 4, ...)

\method{summary}{jmcs}(object, process = c("Longitudinal", "Event"), digits = 4, ...)

\method{summary}{mvjmcs}(object, process = c("Longitudinal", "Event"), digits = 4, ...)
}
\arguments{
\item{object}{an object inheriting from class \code{mvjmcs}.}

\item{digits}{the number of significant digits to use when printing. Default is 4.}

\item{...}{further arguments passed to or from other methods.}

\item{metric}{a list to indicate what metric to summarize}

\item{process}{for which model (i.e., longitudinal model or survival model) to extract the estimated coefficients.}
}
\value{
a list of matrices with conditional probabilities for subjects.

a list of matrices with conditional probabilities for subjects.

a summary of the list of matrices with conditional probabilities for subjects.

a summary of the list of matrices with conditional probabilities for subjects.

A table to summarize the model results.

A table to summarize the model results.

A table to summarize the model results.
}
\description{
Produce result summaries of a joint model fit.

Produce result summaries of a joint model fit.

Produce result summaries of a joint model fit.
}
\seealso{
\code{\link{JMMLSM}, \link{survfitJMMLSM}}

\code{\link{jmcs}, \link{survfitjmcs}}

\code{\link{JMMLSM}, \link{survfitJMMLSM}}

\code{\link{jmcs}, \link{survfitjmcs}}

\code{\link{JMMLSM}}

\code{\link{jmcs}}

\code{\link{mvjmcs}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
