% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.AUCJMMLSM.R,
%   R/summary.ConcordanceJMMLSM.R, R/summary.Concordancejmcs.R,
%   R/summary.DynPredAccjmcs.R, R/summary.JMMLSM.R, R/summary.MAEQJMMLSM.R,
%   R/summary.PEJMMLSM.R, R/summary.jmcs.R, R/summary.mvjmcs.R
\name{summary}
\alias{summary}
\alias{summary.AUCJMMLSM}
\alias{summary.ConcordanceJMMLSM}
\alias{summary.Concordancejmcs}
\alias{summary.DynPredAccjmcs}
\alias{summary.JMMLSM}
\alias{summary.MAEQJMMLSM}
\alias{summary.PEJMMLSM}
\alias{summary.jmcs}
\alias{summary.mvjmcs}
\title{Summaries of evaluation metrics for joint models}
\usage{
\method{summary}{AUCJMMLSM}(object, digits = 4, ...)

\method{summary}{ConcordanceJMMLSM}(object, digits = 4, ...)

\method{summary}{Concordancejmcs}(object, digits = 4, ...)

\method{summary}{DynPredAccjmcs}(
  object,
  metric = c("AUC", "Cindex", "Brier", "MAE", "MAEQ"),
  digits = 4,
  ...
)

\method{summary}{JMMLSM}(object, process = c("longitudinal", "survival"), digits = 4, ...)

\method{summary}{MAEQJMMLSM}(object, digits = 3, ...)

\method{summary}{PEJMMLSM}(object, error = c("MAE", "Brier"), ...)

\method{summary}{jmcs}(object, process = c("Longitudinal", "Event"), digits = 4, ...)

\method{summary}{mvjmcs}(object, process = c("Longitudinal", "Event"), digits = 4, ...)
}
\arguments{
\item{object}{an object inheriting from class \code{mvjmcs}.}

\item{digits}{the number of significant digits to use when printing. Default is 4.}

\item{...}{further arguments passed to or from other methods.}

\item{metric}{a list to indicate what metric to summarize}

\item{process}{for which model (i.e., longitudinal model or survival model) to extract the estimated coefficients.}

\item{error}{a character string that specifies the loss function.}
}
\value{
a list of matrices with conditional probabilities for subjects.

a list of matrices with conditional probabilities for subjects.

a list of matrices with conditional probabilities for subjects.

a summary of the list of matrices with conditional probabilities for subjects.

A table to summarize the model results.

a list of matrices with conditional probabilities for subjects.

a list of matrices with conditional probabilities for subjects.

A table to summarize the model results.

A table to summarize the model results.
}
\description{
Produce result summaries of a joint model fit.

Produce result summaries of a joint model fit.

Produce result summaries of a joint model fit.
}
\seealso{
\code{\link{JMMLSM}, \link{survfitJMMLSM}}

\code{\link{JMMLSM}, \link{survfitJMMLSM}}

\code{\link{jmcs}, \link{survfitjmcs}}

\code{\link{jmcs}, \link{survfitjmcs}}

\code{\link{JMMLSM}}

\code{\link{JMMLSM}, \link{survfitJMMLSM}}

\code{\link{JMMLSM}, \link{survfitJMMLSM}}

\code{\link{jmcs}}

\code{\link{mvjmcs}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
