% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.jmcs.R
\name{summary}
\alias{summary}
\alias{summary.jmcs}
\title{Anova Method for Fitted Joint Models}
\usage{
\method{summary}{jmcs}(object, process = c("longitudinal", "survival"), digits = 4, ...)
}
\arguments{
\item{object}{an object inheriting from class \code{jmcs}.}

\item{process}{for which model (i.e., longitudinal model or survival model) to extract the estimated coefficients.}

\item{digits}{the number of significant digits to use when printing. Default is 4.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A table to summarize the model results.
}
\description{
Produce result summaries of a joint model fit.
}
\seealso{
\code{\link{jmcs}}
}
