% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.JMMLSM.R, R/vcov.jmcs.R, R/vcov.mvjmcs.R
\name{vcov}
\alias{vcov}
\alias{vcov.JMMLSM}
\alias{vcov.jmcs}
\alias{vcov.mvjmcs}
\title{Variance-covariance matrix of the estimated parameters for joint models}
\usage{
\method{vcov}{JMMLSM}(object, ...)

\method{vcov}{jmcs}(object, ...)

\method{vcov}{mvjmcs}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class \code{mvjmcs}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a matrix of variance covariance of all parameter estimates.

a matrix of variance covariance of all parameter estimates.

a matrix of variance covariance of all parameter estimates.
}
\description{
Extract variance-covariance matrix for joint models.

Extract variance-covariance matrix for joint models.

Extract variance-covariance matrix for joint models.
}
\seealso{
\code{\link{JMMLSM}}

\code{\link{jmcs}}

\code{\link{mvjmcs}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
