% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.jmcs.R
\name{vcov}
\alias{vcov}
\alias{vcov.jmcs}
\title{Variance-covariance matrix of the estimated parameters for joint models}
\usage{
\method{vcov}{jmcs}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class \code{jmcs}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a matrix of variance covariance of all parameter estimates.
}
\description{
Extract variance-covariance matrix for joint models.
}
\seealso{
\code{\link{jmcs}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
